Modificações gerais e melhorias
